Initial Catalina support implementation
Due to Catalina's read only system volume (see https://arstechnica.com/gadgets/2019/10/macos-10-15-catalina-the-ars-technica-review/11/#h1), the /server and /sites convenience symlinks on the host are no longer possible.

This host path was hard coded in a few places, and this commit replaces the path with /Volumes/Server. A better solution would be to not depend on host configuration as much, but this allows the dev env to boot successfully on Catalina (and probably earlier).